auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = [
    "Maxim Isaev <isaev_0602@mail.ru>",
    "Ilya Kabeshov <ilua.kabeshov01@mail.ru>"
]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "2.0.27"}
alembic = "1.13.1"
redis = "5.0.1"
async-fastapi-jwt-auth = {extras = ["asymmetric"], version = "0.6.2"}
fastapi = "0.100.1"
uvicorn = "0.27.1"
gunicorn = "21.2.0"
asyncpg = "^0.29.0"
pydantic = "2.6.1"
pydantic-settings = "2.2.1"
passlib = {extras = ["argon2"], version = "1.7.4"}
fakeredis = "^2.21.1"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# -- Flake8 --
[tool.flake8]
max-line-length = 135
exclude = [
    "__submodules__",
    ".git",
    "__samples__",
    ".venv",
    "alembic"
]
max-returns-amount = 10
max-parameters-amount = 25
per-file-ignores = [
    'tests/*:F811',
    '__init__.py:F401'
]

# -- isort --
[tool.isort]
line_length = 135
multi_line_output = 3
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
